Lab 1,  Lab 2
These were the first assignments in the class. 
In the first Lab work (Lab 1), we were required to create a program with the language of our choice (I used Python), that generates two different prime numbers p and q
In the second Lab (Lab 2), We were required to find the Euler's Totient ϕ(n) of the two numbers that have been generated in Lab 1. 
Attached is lab 1 and 2 together with relevant explanations of the code